+++ Setup/UpgradeData.php
@@ -9,10 +9,6 @@ use Magento\Framework\Setup\ModuleContextInterface;
 use Magento\Framework\Setup\ModuleDataSetupInterface;
 use Magento\Framework\Setup\UpgradeDataInterface;
 
-/**
- * Class UpgradeData
- * @package Yireo\Webp2\Setup
- */
 class UpgradeData implements UpgradeDataInterface
 {
     /**