+++ Test/Integration/ImageWithCustomStyleTest.php
@@ -3,12 +3,6 @@ declare(strict_types=1);
 
 namespace Yireo\Webp2\Test\Integration;
 
-use Magento\Framework\View\LayoutInterface;
-
-/**
- * Class ImageWithCustomStyleTest
- * @package Yireo\Webp2\Test\Integration
- */
 class ImageWithCustomStyleTest extends Common
 {
     /**