+++ Test/Utils/ImageProvider.php
@@ -5,10 +5,6 @@ namespace Yireo\Webp2\Test\Utils;
 
 use Magento\Framework\View\Element\Block\ArgumentInterface;
 
-/**
- * Class ImageProvider
- * @package Yireo\Webp2\Test\Utils
- */
 class ImageProvider implements ArgumentInterface
 {
     /**